상수데이터 파일 분리 및 경로 수정
