+++ src/Router.js
@@ -18,7 +18,6 @@ const Router = () => {
         <Route path="/login" element={<Login />} />
         <Route path="/detail" element={<Detail />} />
         <Route path="/signup" element={<SignUp />} />
-        <Route path="/footer" element={<Footer />} />
       </Routes>
       <Footer />
     </BrowserRouter>
