+++ src/pages/Footer/Footer.js
@@ -1,76 +1,24 @@
 import React from 'react';
-import { useNavigate } from 'react-router-dom';
+import { Link } from 'react-router-dom';
+import FOOTER_LIST from './footerList';
+import FOOTER_MEMBER from './footerMember';
 import './Footer.scss';
 
 function Footer() {
-  const navigate = useNavigate();
-  const FOOTER_LIST = [
-    {
-      id: 1,
-      title: 'Product',
-      list: ['Running', 'Soccer', 'Basketball'],
-      url: ['/running', '/soccer', '/basketball'],
-    },
-    {
-      id: 2,
-      title: 'Account',
-      list: ['Login', 'Join', 'Cart'],
-      url: ['/login', '/signup', '/cart'],
-    },
-  ];
-
-  const FOOTER_MEMBER = [
-    {
-      id: 1,
-      name: '김한솔',
-      img: 'images/man.png',
-    },
-    {
-      id: 2,
-      name: '김호준',
-      img: 'images/man1.png',
-    },
-    {
-      id: 3,
-      name: '이명석',
-      img: 'images/man2.png',
-    },
-    {
-      id: 4,
-      name: '조형진',
-      img: 'images/man3.png',
-    },
-    {
-      id: 5,
-      name: '지송현',
-      img: 'images/man4.png',
-    },
-    {
-      id: 6,
-      name: '홍석현',
-      img: 'images/man5.png',
-    },
-    {
-      id: 7,
-      name: '박지영',
-      img: 'images/girl.png',
-    },
-  ];
-
   return (
     <footer>
       <div className="footerInner">
         <div className="footerMenu">
-          {FOOTER_LIST.map(({ id, title, list, url }) => {
+          {FOOTER_LIST.map(({ id, title, list }) => {
             return (
               <div className="footerList" key={id}>
                 <h3>{title}</h3>
                 <ul>
-                  {list.map(text => {
+                  {list.map(({ id, listTitle, path }) => {
                     return (
-                      <li key={text} onClick={() => navigate({ url })}>
-                        {text}
-                      </li>
+                      <Link to={path} key={id}>
+                        <li>{listTitle}</li>
+                      </Link>
                     );
                   })}
                 </ul>
