+++ src/pages/Footer/footerList.js
@@ -1,0 +1,22 @@
+const FOOTER_LIST = [
+  {
+    id: 1,
+    title: 'Product',
+    list: [
+      { id: 1, listTitle: 'Running', path: '/running' },
+      { id: 2, listTitle: 'Soccer', path: '/soccer' },
+      { id: 3, listTitle: 'Basketball', path: '/basketball' },
+    ],
+  },
+  {
+    id: 2,
+    title: 'Account',
+    list: [
+      { id: 1, listTitle: 'Login', path: '/login' },
+      { id: 2, listTitle: 'Join', path: '/signup' },
+      { id: 3, listTitle: 'Cart', path: '/cart' },
+    ],
+  },
+];
+
+export default FOOTER_LIST;
